# src/doc/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "API de Gerenciamento de Fazendas",
      "description": "API para gerenciar produção de leite das fazendas",
      "version": "1.0.0"
    },
    "paths": {
      "/farm": {
        "post": {
          "summary": "Cadastrar uma nova fazenda",
          "tags": ["Fazendas"],
          "security": [{ "bearerAuth": [] }],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FarmInput"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Fazenda cadastrada com sucesso",
              "content": {
                "application/json": {
                  "example": {
                    "message": "Fazenda cadastrada com sucesso",
                    "farm": {
                      "_id": "12345",
                      "nome": "Fazenda Teste",
                      "distancia_km": 10
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Dados inválidos",
              "content": {
                "application/json": {
                  "example": {
                    "errors": [
                      {
                        "msg": "O nome da fazenda é obrigatório"
                      }
                    ]
                  }
                }
              }
            },
            "500": {
              "description": "Erro ao cadastrar a fazenda"
            }
          }
        }
      },
      "/farm/{id}": {
        "get": {
          "summary": "Consultar uma fazenda por ID",
          "tags": ["Fazendas"],
          "security": [{ "bearerAuth": [] }],
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Fazenda encontrada",
              "content": {
                "application/json": {
                  "example": {
                    "_id": "12345",
                    "nome": "Fazenda Teste",
                    "distancia_km": 10
                  }
                }
              }
            },
            "404": {
              "description": "Fazenda não encontrada"
            },
            "500": {
              "description": "Erro ao consultar a fazenda"
            }
          }
        },
        "put": {
          "summary": "Atualizar dados de uma fazenda",
          "tags": ["Fazendas"],
          "security": [{ "bearerAuth": [] }],
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FarmInput"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Fazenda atualizada com sucesso",
              "content": {
                "application/json": {
                  "example": {
                    "message": "Fazenda atualizada com sucesso",
                    "farm": {
                      "_id": "12345",
                      "nome": "Fazenda Teste Atualizada",
                      "distancia_km": 15
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Dados inválidos",
              "content": {
                "application/json": {
                  "example": {
                    "errors": [
                      {
                        "msg": "O nome da fazenda é obrigatório"
                      }
                    ]
                  }
                }
              }
            },
            "404": {
              "description": "Fazenda não encontrada"
            },
            "500": {
              "description": "Erro ao atualizar a fazenda"
            }
          }
        },
        "delete": {
          "summary": "Remover uma fazenda",
          "tags": ["Fazendas"],
          "security": [{ "bearerAuth": [] }],
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Fazenda removida com sucesso",
              "content": {
                "application/json": {
                  "example": {
                    "message": "Fazenda removida com sucesso"
                  }
                }
              }
            },
            "404": {
              "description": "Fazenda não encontrada"
            },
            "500": {
              "description": "Erro ao remover a fazenda"
            }
          }
        }
      },
      "/farmer": {
        "post": {
          "summary": "Cadastrar um novo fazendeiro",
          "tags": ["Fazendeiros"],
          "security": [{ "bearerAuth": [] }],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FarmerInput"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Fazendeiro cadastrado com sucesso",
              "content": {
                "application/json": {
                  "example": {
                    "message": "Fazendeiro cadastrado com sucesso",
                    "farmer": {
                      "_id": "12345",
                      "name": "João da Silva",
                      "email": "joao@example.com",
                      "idFarm": "67890"
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Dados inválidos",
              "content": {
                "application/json": {
                  "example": {
                    "errors": [
                      {
                        "msg": "O nome do fazendeiro é obrigatório"
                      }
                    ]
                  }
                }
              }
            },
            "500": {
              "description": "Erro ao cadastrar o fazendeiro"
            }
          }
        }
      },
      "/farmer/{id}": {
        "get": {
          "summary": "Consultar um fazendeiro por ID",
          "tags": ["Fazendeiros"],
          "security": [{ "bearerAuth": [] }],
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Fazendeiro encontrado",
              "content": {
                "application/json": {
                  "example": {
                    "_id": "12345",
                    "name": "João da Silva",
                    "email": "joao@example.com",
                    "idFarm": "67890"
                  }
                }
              }
            },
            "404": {
              "description": "Fazendeiro não encontrado"
            },
            "500": {
              "description": "Erro ao consultar o fazendeiro"
            }
          }
        },
        "put": {
          "summary": "Atualizar dados de um fazendeiro",
          "tags": ["Fazendeiros"],
          "security": [{ "bearerAuth": [] }],
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FarmerInput"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Fazendeiro atualizado com sucesso",
              "content": {
                "application/json": {
                  "example": {
                    "message": "Fazendeiro atualizado com sucesso",
                    "farmer": {
                      "_id": "12345",
                      "name": "João da Silva Atualizado",
                      "email": "joao@example.com",
                      "idFarm": "67890"
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Dados inválidos",
              "content": {
                "application/json": {
                  "example": {
                    "errors": [
                      {
                        "msg": "O nome do fazendeiro é obrigatório"
                      }
                    ]
                  }
                }
              }
            },
            "404": {
              "description": "Fazendeiro não encontrado"
            },
            "500": {
              "description": "Erro ao atualizar o fazendeiro"
            }
          }
        },
        "delete": {
          "summary": "Remover um fazendeiro",
          "tags": ["Fazendeiros"],
          "security": [{ "bearerAuth": [] }],
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Fazendeiro removido com sucesso",
              "content": {
                "application/json": {
                  "example": {
                    "message": "Fazendeiro removido com sucesso"
                  }
                }
              }
            },
            "404": {
              "description": "Fazendeiro não encontrado"
            },
            "500": {
              "description": "Erro ao remover o fazendeiro"
            }
          }
        }
      },
...
  "/auth/login": {
    "post": {
      "tags": ["Authentication"],
      "summary": "Realizar login e obter tokens",
      "requestBody": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/UserInput"
            }
          }
        },
        "required": true
      },
      "responses": {
        "200": {
          "description": "Tokens gerados com sucesso",
          "content": {
            "application/json": {
              "example": {
                "accessToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
                "refreshToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
              }
            }
          }
        },
        "401": {
          "description": "Credenciais inválidas"
        },
        "500": {
          "description": "Erro ao realizar a autenticação"
        }
      }
    }
  },
  "/auth/refresh": {
    "post": {
      "tags": ["Authentication"],
      "summary": "Gerar um novo token de acesso usando o refresh token",
      "requestBody": {
        "content": {
          "application/json": {
            "schema": {
              "properties": {
                "refreshToken": {
                  "type": "string",
                  "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                }
              },
              "required": ["refreshToken"]
            }
          }
        },
        "required": true
      },
      "responses": {
        "200": {
          "description": "Novo token de acesso gerado com sucesso",
          "content": {
            "application/json": {
              "example": {
                "accessToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
              }
            }
          }
        },
        "403": {
          "description": "Refresh token inválido"
        },
        "500": {
          "description": "Erro ao gerar novo token de acesso"
        }
      }
    }
  }
    },
    "components": {
      "schemas": {
        "FarmInput": {
          "type": "object",
          "properties": {
            "nome": {
              "type": "string"
            },
            "distancia_km": {
              "type": "number"
            }
          },
          "required": ["nome", "distancia_km"]
        },
        "FarmerInput": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string"
              },
              "email": {
                "type": "string"
              },
              "idFarm": {
                "type": "string"
              }
            },
            "required": ["name", "email", "idFarm"]
          },
...
      "UserInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "John Doe"
          },
          "email": {
            "type": "string",
            "example": "john@example.com"
          },
          "password": {
            "type": "string",
            "example": "secret"
          }
        },
        "required": ["name", "email", "password"]
      }
      }
    }
  }